Moved Library functionality
Moved all Library methods over from library app
